refactor(project-structure)!: Restructure codebase into a features-driven directory layout
This commit reorganizes the entire project's file structure to adopt a features-driven architecture. The primary goal is to enhance modularity, improve code organization, and simplify future development and maintenance.

Key changes include:
-   **Scraping Feature (`features/scraping`):** All components related to data scraping, including API routes (`app/api/scrape`), database repositories, Mongoose schemas, scraping services, utilities, and types, have been consolidated under `features/scraping`.
-   **Manual Scheduling Feature (`features/manual_scheduling`):** The scheduling engine classes and services have been moved into `features/manual_scheduling`.
-   **Schema Renaming:** The `lib/db/models` directories have been renamed to `features/*/server/db/schema` to more accurately reflect their purpose as database schemas.
-   **Import Path Updates:** All affected import paths across the application have been updated to reflect the new file locations, ensuring seamless integration within the restructured project.

This refactoring sets a clearer boundary for different functionalities, making the codebase easier to navigate and understand.

BREAKING CHANGE: This commit introduces breaking changes.
