+++ app/api/scrape/degrees/[id]/route.ts
@@ -1,4 +1,5 @@
-import { getDegree } from "@/lib/db/repositories/degree_repository";
+
+import { getDegree } from "@/features/scraping/server/db/repository/degree_repository";
 import { NextRequest, NextResponse } from "next/server";
 
 export const GET = async (
