+++ app/api/scrape/departments/[id]/route.ts
@@ -1,4 +1,4 @@
-import { getDepartment } from "@/lib/db/repositories/department_repository";
+import { getDepartment } from "@/features/scraping/server/db/repository/department_repository";
 import { NextRequest, NextResponse } from "next/server";
 
 export const GET = async (
