+++ features/manual_scheduling/schema/course.ts
