+++ features/manual_scheduling/schema/schedule.ts
