+++ features/scraping/server/db/repository/department_repository.ts
@@ -1,6 +1,6 @@
 import connect from "../db";
 import { DepartmentSchema } from "../db_types";
-import Department from "../models/departments";
+import Department from "../../../features/scraping/server/db/schema/departments";
 
 /**
  * Retrieves a single department by its ID.
