+++ features/scraping/server/db/schema/courses.ts
