+++ features/scraping/server/db/schema/degrees.ts
