+++ features/scraping/types/db_types.ts
